refactor: move gameOver fn to own file
