+++ src/app.js
@@ -3,7 +3,6 @@ import { gameState } from './game-state'
 
 import {
   grid,
-  hiscoreDisplay,
   levelDisplay,
   linesDisplay,
   scoreDisplay,
@@ -19,6 +18,7 @@ import {
   rng,
   setFirstTetromino,
   undrawTetromino,
+  gameOver,
 } from './functions/'
 
 setFirstTetromino(gameState)
@@ -231,47 +231,3 @@ startBtn.addEventListener('click', () => {
     }
   }
 })
-
-/* GAME OVER CONDITION
-checks if any of the squares in the tetromino starting position are taken */
-function gameOver() {
-  if (
-    gameState.currentTetromino.some((index) =>
-      gameState.squares[gameState.currentPosition + index].classList.contains(
-        'taken'
-      )
-    )
-  ) {
-    gameState.isGameOver = true
-
-    for (let i = 0; i < 180; i++) {
-      gameState.squares[i].classList.remove('taken') // removes .taken class from each visible div
-      gameState.squares[i].classList.remove('tetromino') // removes .tetromino class
-    }
-
-    while (grid.firstChild) {
-      grid.removeChild(grid.firstChild) // removes all grid divs from HTML document
-    }
-
-    grid.style.backgroundColor = '#b59aef' // game over background-color
-
-    // TODO: move game over message to its own function
-    const gameOverDiv = document.createElement('div')
-    gameOverDiv.classList.add('text-center', 'game-over')
-    const gameOverText = document.createElement('p')
-    gameOverText.textContent = 'GAME OVER!'
-    gameOverDiv.appendChild(gameOverText)
-    const tryAgainText = document.createElement('p')
-    tryAgainText.textContent = 'PLEASE TRY AGAIN \u2764'
-    gameOverDiv.appendChild(tryAgainText)
-    grid.replaceChildren(gameOverDiv)
-
-    if (gameState.score > gameState.hiscore) {
-      // updates hi score
-      gameState.hiscore = gameState.score * 1
-    }
-
-    hiscoreDisplay.textContent = gameState.hiscore // changes score display
-    clearInterval(gameState.timerId) // clears timer interval for moveDown
-  }
-}
